fix api service file
